Separate support jobs from shield toolkit (#12)
* separate support jobs from shield toolkit

* adjust read me files accordingly
