+++ iot4i-v2/iot4i-shield-toolkit/README.md
@@ -3,10 +3,8 @@ An IBM streams shield development toolkit which contains a set of operators and 
 
 1. source/sink operators that abstract the details of connecting to IBM IoT platform.
 2. source/sink operators that abstract the details of connecting to IBM message hub service.
-3. CheckActivation operator which can be used to connect with IoT4i API and verify whether the shield is activated or not for a particular user.
-4. Transformer operator which annotate the raw sensor events with user information.  
-5. Action sink operator which calls IoT4i API to create a hazard and invoke actions.
-6. Enforce and validate the hazards generated by shield jobs.
+3. CheckActivation operator which can be used to connect with IoT4i API and verify whether the shield is activated or not for a particular user. 
+4. Enforce and validate the hazards generated by shield jobs.
 
 ### Structure
 
@@ -18,14 +16,7 @@ This is the stage where the events are prepare for the shield jobs. The followin
 
 **Common Operators**:
  
- 1. [TransformerOp](./impl/java/src/com/ibm/iot4i/events/common/TransformerOp.java):
- A java primitive operator to annotated raw events with user information. This operator adds also unique eventIds to the raw events. The operator accepts the following input params:
- 
- `deviceIdAttributeNames`:  The name of the attribute in the event payload that contains the physical device Id. The operator will check by default the deviceId of the IoTP topic, if not deviceId is found, it will search in the payload based on the attribute name provided here.
- 
- The operator requires also `apiURL`, `apiToken`, and `tenantId` but this will be provided by the IoT4i API when the job is submitted.
- 
- 2. [CheckActivationOp](./impl/java/src/com/ibm/iot4i/events/common/CheckActivationOp.java):
+ [CheckActivationOp](./impl/java/src/com/ibm/iot4i/events/common/CheckActivationOp.java):
  
  A java primitive operator to check if the user has activated his shield or not. If the user didn't active his shield, the events will be filtered out. The operator requires the following input params:
  
@@ -37,11 +28,7 @@ This is the stage where the events are prepare for the shield jobs. The followin
  
  A source operator that abstract the details of connecting to IBM IoT platform. This operator has no input and generate an output with type `stream<rstring key, rstring message>`. The key contains the IoTP topic and the message contains the payload.
 
- 2. [Transformer](./com.ibm.iot4i.events.IoTP/Transformer.spl)
- 
- An operator that receives raw events and transform them using the [TransformerOp](./impl/java/src/com/ibm/iot4i/events/common/TransformerOp.java). The operator refresh the cache of the TransformerOp by listening to operational notifications.
- 
- 3. [CheckActivation](./com.ibm.iot4i.events.IoTP/CheckActivation.spl)
+ 2. [CheckActivation](./com.ibm.iot4i.events.IoTP/CheckActivation.spl)
  
  An operator that receives transformed events and check filter events based on the [CheckActivationOp](./impl/java/src/com/ibm/iot4i/events/common/CheckActivationOp.java).
  
@@ -56,10 +43,6 @@ This is the stage where the events are prepare for the shield jobs. The followin
   
   A source operator that listen to the annotated events which are generated by the transformer job. This operator uses the Source operator internally but listens to different topic and make sure the userId is added to the events. 
  
-  2. [Transformer](./com.ibm.iot4i.events.MH/Transformer.spl)
-  
-  An operator that receives raw events and transform them using the [TransformerOp](./impl/java/src/com/ibm/iot4i/events/common/TransformerOp.java). The operator refresh the cache of the TransformerOp by listening to operational notifications.
-  
   3. [CheckActivation](./com.ibm.iot4i.events.MH/CheckActivation.spl)
   
   An operator that receives transformed events and check filter events based on the [CheckActivationOp](./impl/java/src/com/ibm/iot4i/events/common/CheckActivationOp.java).
@@ -71,13 +54,9 @@ In this stage, the hazard generated by the shield jobs are prepared for submissi
 
 **Common Operators**:
 
-1. [ActionSink](./com.ibm.iot4i.hazards.common/ActionSink.spl)
+[PrepareHazards](./com.ibm.iot4i.hazards.common/PrepareHazards.spl)
 
-An operator that receives hazards and call the create hazard API of IoT4i.
-
-2. [PrepareHazards](./com.ibm.iot4i.hazards.common/PrepareHazards.spl)
-
-An operator that prepare the hazard for all sinks. It parse the hazard to string, adds shieldId and make sure that the userId exists. 
+An operator that prepare the hazard for sinks. It parse the hazard to string, adds shieldId and make sure that the userId exists. 
 
 **IBM IoT Platform Operators**:
 
@@ -87,7 +66,7 @@ An operator that receives hazards and publish them to IoTP.
 
 2. [SubmitHazards](./com.ibm.iot4i.hazards.IoTP/SubmitHazards.spl)
 
-An operator that receives hazards, prepare them, then calls the IoTP and Action sinks. 
+An operator that receives hazards, prepare them, then calls sinks. 
 
 **Message Hub Operators**:
