+++ iot4i-v2/shield-templates/toolkit.xml
@@ -4,29 +4,16 @@
   <toolkit name="ShieldTemplates" requiredProductVersion="4.2.1.1" version="1.0.0">
     <description/>
     <uriTable>
-      <uri index="2" value="com.ibm.iot4i.shields.templates/ShieldTemplateWithIoTP.spl"/>
-      <uri index="1" value="com.ibm.iot4i.shields.templates/ShieldTemplateWithMH.spl"/>
+      <uri index="1" value="com.ibm.iot4i.shields.templates/ShieldTemplate.spl"/>
     </uriTable>
     <namespace name="com.ibm.iot4i.shields.templates">
-      <compositeOp column="11" line="7" name="ShieldTemplateWithMH" potentialMain="true" uriIndex="1">
+      <compositeOp column="11" line="7" name="ShieldTemplate" potentialMain="true" uriIndex="1">
         <parameter defaultValue="getSubmissionTimeValue(&quot;KAFKA_BROKERS_SASL&quot;)" metaType="Expression" name="KAFKA_BROKERS_SASL" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;USERNAME&quot;)" metaType="Expression" name="USERNAME" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;PASSWORD&quot;)" metaType="Expression" name="PASSWORD" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;SOURCE_TOPIC&quot;)" metaType="Expression" name="SOURCE_TOPIC" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;NOTIFICATION_TOPIC&quot;)" metaType="Expression" name="NOTIFICATION_TOPIC" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;SINK_TOPIC&quot;)" metaType="Expression" name="SINK_TOPIC" optional="true" type="&lt;rstring>"/>
-        <parameter defaultValue="getSubmissionTimeValue(&quot;API_URL&quot;) + &quot;/&quot;" metaType="Expression" name="API_URL" optional="true" type="&lt;rstring>"/>
-        <parameter defaultValue="getSubmissionTimeValue(&quot;API_TOKEN&quot;)" metaType="Expression" name="API_TOKEN" optional="true" type="&lt;rstring>"/>
-        <parameter defaultValue="getSubmissionTimeValue(&quot;SHIELD_ID&quot;)" metaType="Expression" name="SHIELD_ID" optional="true" type="&lt;rstring>"/>
-        <parameter defaultValue="getSubmissionTimeValue(&quot;TENANT_ID&quot;)" metaType="Expression" name="TENANT_ID" optional="true" type="&lt;rstring>"/>
-      </compositeOp>
-      <compositeOp column="11" line="7" name="ShieldTemplateWithIoTP" potentialMain="true" uriIndex="2">
-        <parameter defaultValue="getSubmissionTimeValue(&quot;ORG&quot;)" metaType="Expression" name="ORG" optional="true" type="&lt;rstring>"/>
-        <parameter defaultValue="getSubmissionTimeValue(&quot;AUTH_KEY&quot;)" metaType="Expression" name="AUTH_KEY" optional="true" type="&lt;rstring>"/>
-        <parameter defaultValue="getSubmissionTimeValue(&quot;AUTH_TOKEN&quot;)" metaType="Expression" name="AUTH_TOKEN" optional="true" type="&lt;rstring>"/>
-        <parameter defaultValue="(int32)getSubmissionTimeValue(&quot;QOS&quot;)" metaType="Expression" name="QOS" optional="true" type="&lt;int32>"/>
-        <parameter defaultValue="&quot;iot-2/type/+/id/+/evt/+/fmt/+&quot;" metaType="Expression" name="SOURCE_TOPIC" optional="true" type="&lt;rstring>"/>
-        <parameter defaultValue="(rstring)jobID() + &quot;_evt&quot;" metaType="Expression" name="SOURCE_APP_ID" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;API_URL&quot;) + &quot;/&quot;" metaType="Expression" name="API_URL" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;API_TOKEN&quot;)" metaType="Expression" name="API_TOKEN" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;SHIELD_ID&quot;)" metaType="Expression" name="SHIELD_ID" optional="true" type="&lt;rstring>"/>
