+++ iot4i-v2/support-jobs/ActionJob/README.md
@@ -3,6 +3,22 @@
 
 This job listens to all hazards, prepare hazards then calls the create hazard API to create the hazard document and invoke the external action. 
 
- Inline-style: 
  ![alt text](./images/ActionJob.png "Action Job")
  
+ 
+ This job uses the following operator:
+ 
+ [ActionSink](./com.ibm.iot4i.actionjob/ActionSink.spl)
+ 
+ 
+ Expected Input Hazard Message:
+ 
+```json
+ {
+ 	"userId": "",
+ 	"actionParams": {},
+ 	"rawEvents": []
+ }
+````
+
+
