+++ iot4i-v2/support-jobs/ActionJob/toolkit.xml
@@ -4,10 +4,19 @@
   <toolkit name="ActionJob" requiredProductVersion="4.2.1.1" version="1.0.0">
     <description/>
     <uriTable>
-      <uri index="1" value="com.ibm.iot4i.actionjob/Main.spl"/>
+      <uri index="3" value="com.ibm.iot4i.actionjob/native.function/javaFunction.xml"/>
+      <uri index="1" value="com.ibm.iot4i.actionjob/ActionSink.spl"/>
+      <uri index="2" value="com.ibm.iot4i.actionjob/Main.spl"/>
     </uriTable>
     <namespace name="com.ibm.iot4i.actionjob">
-      <compositeOp column="11" line="6" name="Main" potentialMain="true" uriIndex="1">
+      <compositeOp column="18" line="8" name="ActionSink" potentialMain="false" public="true" uriIndex="1">
+        <parameter metaType="Expression" name="API_URL" optional="false" type="&lt;rstring>"/>
+        <parameter metaType="Expression" name="API_TOKEN" optional="false" type="&lt;rstring>"/>
+        <parameter defaultValue="10" metaType="Expression" name="RETRY_COUNT" optional="true" type="&lt;int32>"/>
+        <parameter metaType="Expression" name="TENANT_ID" optional="false" type="&lt;rstring>"/>
+        <inputPort name="PreparedHazardStream" portIndex="0" type="stream&lt;rstring data>"/>
+      </compositeOp>
+      <compositeOp column="11" line="6" name="Main" potentialMain="true" uriIndex="2">
         <parameter defaultValue="getSubmissionTimeValue(&quot;KAFKA_BROKERS_SASL&quot;)" metaType="Expression" name="KAFKA_BROKERS_SASL" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;USERNAME&quot;)" metaType="Expression" name="USERNAME" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;PASSWORD&quot;)" metaType="Expression" name="PASSWORD" optional="true" type="&lt;rstring>"/>
@@ -16,6 +25,11 @@
         <parameter defaultValue="getSubmissionTimeValue(&quot;API_TOKEN&quot;)" metaType="Expression" name="API_TOKEN" optional="true" type="&lt;rstring>"/>
         <parameter defaultValue="getSubmissionTimeValue(&quot;TENANT_ID&quot;)" metaType="Expression" name="TENANT_ID" optional="true" type="&lt;rstring>"/>
       </compositeOp>
+      <function modelUriIndex="3" name="isHttps" native="true" public="true" returnType="boolean" uriIndex="0">
+        <prototype>&lt;string T&gt; public boolean isHttps(T apiURL)</prototype>
+        <parameter name="apiURL" type="T"/>
+      </function>
+      <type column="6" line="6" name="BodySchema" static="true" type="rstring data" uriIndex="1"/>
     </namespace>
     <dependency>
       <common:name>iot4i-shield-toolkit</common:name>
