+++ iot4i-v2/support-jobs/TransformerJob/com.ibm.iot4i.transformerjob/TransformerOp/TransformerOp.xml
@@ -1,0 +1,71 @@
+<?xml version="1.0" encoding="UTF-8"?><operatorModel xmlns="http://www.ibm.com/xmlns/prod/streams/spl/operator" xmlns:cmn="http://www.ibm.com/xmlns/prod/streams/spl/common" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.ibm.com/xmlns/prod/streams/spl/operator operatorModel.xsd">
+  <!--DO NOT EDIT THIS FILE - it is machine generated-->
+  <javaOperatorModel>
+    <!--Generated from com.ibm.iot4i.transformerjob.TransformerOp in impl/java/bin/com/ibm/iot4i/transformerjob/TransformerOp$StreamsModel.class at Wed Aug 02 11:27:11 EDT 2017-->
+    <context>
+      <description>Java Operator Transformer</description>
+      <metrics/>
+      <executionSettings>
+        <className>com.ibm.iot4i.transformerjob.TransformerOp</className>
+      </executionSettings>
+      <libraryDependencies>
+        <library>
+          <cmn:description>Operator class library</cmn:description>
+          <cmn:managedLibrary>
+            <cmn:libPath>../../impl/java/bin</cmn:libPath>
+            <cmn:libPath>../../impl/lib/*</cmn:libPath>
+          </cmn:managedLibrary>
+        </library>
+      </libraryDependencies>
+    </context>
+    <parameters>
+      <parameter>
+        <name>apiToken</name>
+        <description/>
+        <optional>false</optional>
+        <type>rstring</type>
+        <cardinality>1</cardinality>
+      </parameter>
+      <parameter>
+        <name>apiURL</name>
+        <description/>
+        <optional>false</optional>
+        <type>rstring</type>
+        <cardinality>1</cardinality>
+      </parameter>
+      <parameter>
+        <name>deviceIdAttributeNames</name>
+        <description/>
+        <optional>false</optional>
+        <type>rstring</type>
+        <cardinality>1</cardinality>
+      </parameter>
+      <parameter>
+        <name>tenantId</name>
+        <description/>
+        <optional>false</optional>
+        <type>rstring</type>
+        <cardinality>1</cardinality>
+      </parameter>
+    </parameters>
+    <inputPorts>
+      <inputPortSet>
+        <description>Port that ingests tuples</description>
+        <windowingDescription/>
+        <windowingMode>NonWindowed</windowingMode>
+        <windowPunctuationInputMode>Oblivious</windowPunctuationInputMode>
+        <controlPort>false</controlPort>
+        <cardinality>1</cardinality>
+        <optional>false</optional>
+      </inputPortSet>
+    </inputPorts>
+    <outputPorts>
+      <outputPortSet>
+        <description>Port that produces tuples</description>
+        <windowPunctuationOutputMode>Generating</windowPunctuationOutputMode>
+        <cardinality>1</cardinality>
+        <optional>false</optional>
+      </outputPortSet>
+    </outputPorts>
+  </javaOperatorModel>
+</operatorModel>
